Update tests to ensure comparison is using double types

+++ tests/dimMtimesTest.m
@@ -13,25 +13,25 @@ tol = 1e-5;
 val = 2 * oneInch;
 expected = DimInch(2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = 2 * oneMillimeter;
 expected = DimMillimeter(2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = 0 * oneMillimeter;
 expected = DimMillimeter(0);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = -2 * oneMillimeter;
 expected = DimMillimeter(-2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 %% Test 2: mtimes of scalar on RHS
@@ -39,22 +39,22 @@ assert(abs(val - expected) < tol);
 val = oneInch * 2;
 expected = DimInch(2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = oneMillimeter * 2;
 expected = DimMillimeter(2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = oneMillimeter * 0;
 expected = DimMillimeter(0);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = oneMillimeter * -2;
 expected = DimMillimeter(-2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
