+++ tests/dimUplusTest.m
@@ -16,23 +16,23 @@ tol = 1e-5;
 val = +negOneInch;
 expected = DimInch(1);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = +negTwoInches;
 expected = DimInch(2);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = +negOneMillimeter;
 expected = DimMillimeter(1);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
 
 
 val = +neg254Millimeters;
 expected = DimMillimeter(254);
 assert(strcmp(class(val), class(expected)));
-assert(abs(val - expected) < tol);
+assert(abs(double(val) - double(expected)) < tol);
